test for valid migration job instantiations
